split session.go to {transaction,data_query,statement,keep_in_cache}.go
