+++ internal/table/data_query.go
@@ -1,0 +1,47 @@
+package table
+
+import "github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Table"
+
+type dataQuery struct {
+	query    Ydb_Table.Query
+	queryID  Ydb_Table.Query_Id
+	queryYQL Ydb_Table.Query_YqlText
+}
+
+func (q *dataQuery) String() string {
+	var emptyID Ydb_Table.Query_Id
+	if q.queryID == emptyID {
+		return q.queryYQL.YqlText
+	}
+	return q.queryID.Id
+}
+
+func (q *dataQuery) ID() string {
+	return q.queryID.Id
+}
+
+func (q *dataQuery) YQL() string {
+	return q.queryYQL.YqlText
+}
+
+func (q *dataQuery) initFromText(s string) {
+	q.queryID = Ydb_Table.Query_Id{} // Reset id field.
+	q.queryYQL.YqlText = s
+	q.query.Query = &q.queryYQL
+}
+
+func (q *dataQuery) initPrepared(id string) {
+	q.queryYQL = Ydb_Table.Query_YqlText{} // Reset yql field.
+	q.queryID.Id = id
+	q.query.Query = &q.queryID
+}
+
+func (q *dataQuery) initPreparedText(s, id string) {
+	q.queryYQL = Ydb_Table.Query_YqlText{} // Reset yql field.
+	q.queryYQL.YqlText = s
+
+	q.queryID = Ydb_Table.Query_Id{} // Reset id field.
+	q.queryID.Id = id
+
+	q.query.Query = &q.queryID // Prefer preared query.
+}
