+++ internal/table/session_test.go
@@ -7,14 +7,11 @@ import (
 	"testing"
 	"time"
 
-	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/anypb"
-
 	"github.com/ydb-platform/ydb-go-genproto/Ydb_Table_V1"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
-	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Operations"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Scheme"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Table"
+	"google.golang.org/protobuf/proto"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/operation"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/value"
@@ -289,7 +286,7 @@ func TestSessionOperationModeOnExecuteDataQuery(t *testing.T) {
 		{
 			method: testutil.TableCommitTransaction,
 			do: func(t *testing.T, ctx context.Context, c *client) {
-				tx := &Transaction{
+				tx := &transaction{
 					s: &session{
 						tableService: Ydb_Table_V1.NewTableServiceClient(c.cc),
 					},
@@ -301,7 +298,7 @@ func TestSessionOperationModeOnExecuteDataQuery(t *testing.T) {
 		{
 			method: testutil.TableRollbackTransaction,
 			do: func(t *testing.T, ctx context.Context, c *client) {
-				tx := &Transaction{
+				tx := &transaction{
 					s: &session{
 						tableService: Ydb_Table_V1.NewTableServiceClient(c.cc),
 					},
@@ -392,215 +389,5 @@ func TestSessionOperationModeOnExecuteDataQuery(t *testing.T) {
 				}
 			},
 		)
-	}
-}
-
-func TestQueryCachePolicyKeepInCache(t *testing.T) {
-	for _, test := range [...]struct {
-		name                   string
-		queryCachePolicyOption []options.QueryCachePolicyOption
-	}{
-		{
-			name: "with server cache",
-			queryCachePolicyOption: []options.QueryCachePolicyOption{
-				options.WithQueryCachePolicyKeepInCache(),
-			},
-		},
-		{
-			name:                   "no server cache",
-			queryCachePolicyOption: []options.QueryCachePolicyOption{},
-		},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			b := StubBuilder{
-				T: t,
-				cc: testutil.NewDB(
-					testutil.WithInvokeHandlers(
-						testutil.InvokeHandlers{
-							// nolint:unparam
-							testutil.TableExecuteDataQuery: func(request interface{}) (proto.Message, error) {
-								r, ok := request.(*Ydb_Table.ExecuteDataQueryRequest)
-								if !ok {
-									t.Fatalf("cannot cast request '%T' to *Ydb_Table.ExecuteDataQueryRequest", request)
-								}
-								if len(test.queryCachePolicyOption) > 0 {
-									if !r.QueryCachePolicy.GetKeepInCache() {
-										t.Fatalf("keep-in-cache policy must be true, got: %v", r.QueryCachePolicy.GetKeepInCache())
-									}
-								} else {
-									if r.QueryCachePolicy.GetKeepInCache() {
-										t.Fatalf("keep-in-cache policy must be false, got: %v", r.QueryCachePolicy.GetKeepInCache())
-									}
-								}
-								return &Ydb_Table.ExecuteQueryResult{
-									TxMeta: &Ydb_Table.TransactionMeta{
-										Id: "",
-									},
-								}, nil
-							},
-							// nolint:unparam
-							testutil.TableCreateSession: func(interface{}) (proto.Message, error) {
-								return &Ydb_Table.CreateSessionResult{
-									SessionId: testutil.SessionID(),
-								}, nil
-							},
-						},
-					),
-				),
-			}
-			s, err := b.createSession(context.Background())
-			if err != nil {
-				t.Fatal(err)
-			}
-			_, _, err = s.Execute(
-				context.Background(), table.TxControl(
-					table.BeginTx(
-						table.WithOnlineReadOnly(),
-					),
-					table.CommitTx(),
-				),
-				"SELECT 1",
-				table.NewQueryParameters(),
-				options.WithQueryCachePolicy(test.queryCachePolicyOption...),
-			)
-			if err != nil {
-				t.Fatal(err)
-			}
-		})
-	}
-}
-
-func TestTxSkipRollbackForCommitted(t *testing.T) {
-	var (
-		begin    = 0
-		commit   = 0
-		rollback = 0
-	)
-	b := StubBuilder{
-		T: t,
-		cc: testutil.NewDB(
-			testutil.WithInvokeHandlers(
-				testutil.InvokeHandlers{
-					testutil.TableBeginTransaction: func(request interface{}) (proto.Message, error) {
-						_, ok := request.(*Ydb_Table.BeginTransactionRequest)
-						if !ok {
-							t.Fatalf("cannot cast request '%T' to *Ydb_Table.BeginTransactionRequest", request)
-						}
-						result, err := anypb.New(
-							&Ydb_Table.BeginTransactionResult{
-								TxMeta: &Ydb_Table.TransactionMeta{
-									Id: "",
-								},
-							},
-						)
-						if err != nil {
-							return nil, err
-						}
-						begin++
-						return &Ydb_Table.BeginTransactionResponse{
-							Operation: &Ydb_Operations.Operation{
-								Ready:  true,
-								Status: Ydb.StatusIds_SUCCESS,
-								Result: result,
-							},
-						}, nil
-					},
-					testutil.TableCommitTransaction: func(request interface{}) (proto.Message, error) {
-						_, ok := request.(*Ydb_Table.CommitTransactionRequest)
-						if !ok {
-							t.Fatalf("cannot cast request '%T' to *Ydb_Table.CommitTransactionRequest", request)
-						}
-						result, err := anypb.New(
-							&Ydb_Table.CommitTransactionResult{},
-						)
-						if err != nil {
-							return nil, err
-						}
-						commit++
-						return &Ydb_Table.CommitTransactionResponse{
-							Operation: &Ydb_Operations.Operation{
-								Ready:  true,
-								Status: Ydb.StatusIds_SUCCESS,
-								Result: result,
-							},
-						}, nil
-					},
-					// nolint:unparam
-					testutil.TableRollbackTransaction: func(request interface{}) (proto.Message, error) {
-						_, ok := request.(*Ydb_Table.RollbackTransactionRequest)
-						if !ok {
-							t.Fatalf("cannot cast request '%T' to *Ydb_Table.RollbackTransactionRequest", request)
-						}
-						rollback++
-						return &Ydb_Table.RollbackTransactionResponse{
-							Operation: &Ydb_Operations.Operation{
-								Ready:  true,
-								Status: Ydb.StatusIds_SUCCESS,
-							},
-						}, nil
-					},
-					// nolint:unparam
-					testutil.TableCreateSession: func(interface{}) (proto.Message, error) {
-						return &Ydb_Table.CreateSessionResult{
-							SessionId: testutil.SessionID(),
-						}, nil
-					},
-				},
-			),
-		),
-	}
-	s, err := b.createSession(context.Background())
-	if err != nil {
-		t.Fatal(err)
-	}
-	{
-		x, err := s.BeginTransaction(context.Background(), table.TxSettings())
-		if err != nil {
-			t.Fatal(err)
-		}
-		if begin != 1 {
-			t.Fatalf("unexpected begin: %d", begin)
-		}
-		_, err = x.CommitTx(context.Background())
-		if err != nil {
-			t.Fatal(err)
-		}
-		if commit != 1 {
-			t.Fatalf("unexpected commit: %d", begin)
-		}
-		_, _ = x.CommitTx(context.Background())
-		if commit != 1 {
-			t.Fatalf("unexpected commit: %d", begin)
-		}
-		err = x.Rollback(context.Background())
-		if err != nil {
-			t.Fatal(err)
-		}
-		if rollback != 0 {
-			t.Fatalf("unexpected rollback: %d", begin)
-		}
-	}
-	{
-		x, err := s.BeginTransaction(context.Background(), table.TxSettings())
-		if err != nil {
-			t.Fatal(err)
-		}
-		if begin != 2 {
-			t.Fatalf("unexpected begin: %d", begin)
-		}
-		err = x.Rollback(context.Background())
-		if err != nil {
-			t.Fatal(err)
-		}
-		if rollback != 1 {
-			t.Fatalf("unexpected rollback: %d", begin)
-		}
-		_, err = x.CommitTx(context.Background())
-		if err != nil {
-			t.Fatal(err)
-		}
-		if commit != 2 {
-			t.Fatalf("unexpected commit: %d", begin)
-		}
 	}
 }
